+++ cmake/DetectHeaders.cmake
@@ -47,6 +47,39 @@ endif()
 
 # -------------------------------------------------------------------------------------------------- #
 check_c_source_compiles("
+  #include <sys/mman.h>
+  int main( void ) {
+     return 0;
+  }" LIBAKRYPT_HAVE_SYSMMAN )
+
+if( LIBAKRYPT_HAVE_SYSMMAN )
+    set( CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLIBAKRYPT_HAVE_SYSMMAN_H" )
+endif()
+
+# -------------------------------------------------------------------------------------------------- #
+check_c_source_compiles("
+  #include <sys/stat.h>
+  int main( void ) {
+     return 0;
+  }" LIBAKRYPT_HAVE_SYSSTAT )
+
+if( LIBAKRYPT_HAVE_SYSSTAT )
+    set( CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLIBAKRYPT_HAVE_SYSSTAT_H" )
+endif()
+
+# -------------------------------------------------------------------------------------------------- #
+check_c_source_compiles("
+  #include <errno.h>
+  int main( void ) {
+     return 0;
+  }" LIBAKRYPT_HAVE_ERRNO )
+
+if( LIBAKRYPT_HAVE_ERRNO )
+    set( CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DLIBAKRYPT_HAVE_ERRNO_H" )
+endif()
+
+# -------------------------------------------------------------------------------------------------- #
+check_c_source_compiles("
   #include <getopt.h>
   int main( void ) {
      return 0;