+++ source/libakrypt.h
@@ -333,6 +333,8 @@
  dll_export size_t ak_hash_get_block_size( ak_handle );
 /*! \brief Хеширование заданной области памяти. */
  dll_export ak_buffer ak_hash_ptr( ak_handle , const ak_pointer , const size_t , ak_pointer );
+/*! \brief Хеширование заданного файла. */
+ dll_export ak_buffer ak_hash_file( ak_handle , const char* , ak_pointer );
 
 /* ----------------------------------------------------------------------------------------------- */
 /*! \brief Поиск OID по типу криптографического механизма. */